bugfixes & added feature
After video 9 I started modeling the "reopen" flow. Here I encountered and fixed a bug in the expected event configuration in the Test Case. After this I finished the reopen flow.
Then I continued modeling the "CreateTask" flow.
I fixed a bug in the "emit-event" processor configuration. The select list was not initiated a lot of times.
I fixed the bug in the expected state modeling, I'm still testing. But it looks good.
I had a similar bug in the initial state modeling.
I added a new feature in the model wizard, when creating a domain event you can now target a sub collection explicitly instead of the document as a whole.
After these alterations to TP I finished the create task flow.
 The update task was modeled without any issues.
The complete task flow was modeled without any issues.